Update BpmnXMLUtil.java
So that callers of parseChildElements() can override the behavior of an existing child parser. Previously it was not allowed because childParser.putAll(genericChildParserMap) overwrites any custom child parser with the same element name as a generic child parser. Also removes the unwanted side effect of childParsers being filled with all genericChildParsers after calling this method.